cmake: hide libelf behind loader.h impl
Extract elf_loader into its own header/.cc so all libelf types and
includes live in the .cc only; loader.h delegates elf_load() to
gs::elf_loader. Move the libelf include dir to PRIVATE. This lets
libelf become a PRIVATE dependency, needed for find_package(qbox).
