+++ systemc-components/common/include/loader.h
@@ -20,13 +20,11 @@
 #include <tlm_sockets_buswidth.h>
 
 #include "onmethod.h"
+#include "elf_loader.h"
 
 #include <cinttypes>
-#include <fcntl.h>
-#include <libelf.h>
 #include <list>
 #include <string>
-#include <unistd.h>
 #include <vector>
 #include <limits>
 #include <filesystem>
@@ -499,220 +497,8 @@ public:
 
     void elf_load(const std::string& path)
     {
-        elf_reader(path, [&](uint64_t addr, uint8_t* data, uint64_t len) -> void { send(addr, data, len); });
+        gs::elf_loader(path, [&](uint64_t addr, uint8_t* data, uint64_t len) -> void { send(addr, data, len); });
     }
-
-    /* Elf reader helper class */
-private:
-    /* elf loading */
-    typedef enum _endianess {
-        ENDIAN_UNKNOWN = 0,
-        ENDIAN_LITTLE = 1,
-        ENDIAN_BIG = 2,
-    } endianess;
-
-    struct elf_segment {
-        const uint64_t virt;
-        const uint64_t phys;
-        const uint64_t size;
-        const uint64_t filesz;
-        const uint64_t offset;
-        const bool r, w, x;
-    };
-
-    class elf_reader
-    {
-    private:
-        std::vector<struct elf_segment> m_segments;
-
-        std::function<void(uint64_t, uint8_t*, uint64_t)> m_send;
-
-        std::string m_filename;
-        int m_fd;
-        uint64_t m_entry;
-        uint64_t m_machine;
-        endianess m_endian;
-
-    public:
-        uint64_t entry() const { return m_entry; }
-        uint64_t machine() const { return m_machine; }
-
-        endianess endian() const { return m_endian; }
-        bool is_little_endian() const { return m_endian == ENDIAN_LITTLE; }
-        bool is_big_endian() const { return m_endian == ENDIAN_BIG; }
-
-        const char* filename() const { return m_filename.c_str(); }
-
-        const std::vector<elf_segment>& segments() const { return m_segments; }
-
-        elf_reader(const elf_reader&) = delete;
-
-        struct elf32_traits {
-            typedef Elf32_Ehdr Elf_Ehdr;
-            typedef Elf32_Phdr Elf_Phdr;
-            typedef Elf32_Shdr Elf_Shdr;
-            typedef Elf32_Sym Elf_Sym;
-
-            static Elf_Ehdr* elf_getehdr(Elf* elf) { return elf32_getehdr(elf); }
-
-            static Elf_Phdr* elf_getphdr(Elf* elf) { return elf32_getphdr(elf); }
-
-            static Elf_Shdr* elf_getshdr(Elf_Scn* scn) { return elf32_getshdr(scn); }
-        };
-
-        struct elf64_traits {
-            typedef Elf64_Ehdr Elf_Ehdr;
-            typedef Elf64_Phdr Elf_Phdr;
-            typedef Elf64_Shdr Elf_Shdr;
-            typedef Elf64_Sym Elf_Sym;
-
-            static Elf_Ehdr* elf_getehdr(Elf* elf) { return elf64_getehdr(elf); }
-
-            static Elf_Phdr* elf_getphdr(Elf* elf) { return elf64_getphdr(elf); }
-
-            static Elf_Shdr* elf_getshdr(Elf_Scn* scn) { return elf64_getshdr(scn); }
-        };
-
-        static endianess elf_endianess(Elf* elf)
-        {
-            const char* ident = elf_getident(elf, nullptr);
-            if (ident == nullptr) return ENDIAN_UNKNOWN;
-
-            switch (ident[EI_DATA]) {
-            case ELFDATA2LSB:
-                return ENDIAN_LITTLE;
-            case ELFDATA2MSB:
-                return ENDIAN_BIG;
-            default:
-                return ENDIAN_UNKNOWN;
-            }
-        }
-
-        template <typename T>
-        static std::vector<elf_segment> elf_segments(Elf* elf)
-        {
-            size_t count = 0;
-            int err = elf_getphdrnum(elf, &count);
-            if (err) SCP_FATAL("elf_reader") << "elf_begin failed: " << elf_errmsg(err);
-
-            std::vector<elf_segment> segments;
-            typename T::Elf_Phdr* hdr = T::elf_getphdr(elf);
-            for (uint64_t i = 0; i < count; i++, hdr++) {
-                if (hdr->p_type == PT_LOAD) {
-                    bool r = hdr->p_flags & PF_R;
-                    bool w = hdr->p_flags & PF_W;
-                    bool x = hdr->p_flags & PF_X;
-                    segments.push_back(
-                        { hdr->p_vaddr, hdr->p_paddr, hdr->p_memsz, hdr->p_filesz, hdr->p_offset, r, w, x });
-                }
-            }
-
-            return segments;
-        }
-
-        template <typename T, typename ELF>
-        void read_sections(ELF* elf)
-        {
-            Elf_Scn* scn = nullptr;
-            while ((scn = elf_nextscn(elf, scn)) != nullptr) {
-                typename T::Elf_Shdr* shdr = T::elf_getshdr(scn);
-                if (shdr->sh_type != SHT_SYMTAB) continue;
-
-                Elf_Data* data = elf_getdata(scn, nullptr);
-                size_t num_symbols = shdr->sh_size / shdr->sh_entsize;
-                typename T::Elf_Sym* syms = (typename T::Elf_Sym*)(data->d_buf);
-
-                for (size_t i = 0; i < num_symbols; i++) {
-                    if (syms[i].st_size == 0) continue;
-
-                    char* name = elf_strptr(elf, shdr->sh_link, syms[i].st_name);
-                    if (name == nullptr || strlen(name) == 0) continue;
-                }
-            }
-        }
-
-        uint64_t to_phys(uint64_t virt) const
-        {
-            for (auto& seg : m_segments) {
-                if ((virt >= seg.virt) && virt < (seg.virt + seg.size)) return seg.phys + virt - seg.virt;
-            }
-
-            return virt;
-        }
-
-        elf_reader(const std::string& path, std::function<void(uint64_t, uint8_t*, uint64_t)> _send)
-            : m_send(_send), m_filename(path), m_fd(-1), m_entry(0), m_machine(0), m_endian(ENDIAN_UNKNOWN)
-        {
-            if (elf_version(EV_CURRENT) == EV_NONE) SCP_FATAL("elf_reader") << "failed to read libelf version";
-
-#ifdef _WIN32
-            m_fd = open(filename(), O_RDONLY | O_BINARY, 0);
-#else
-            m_fd = open(filename(), O_RDONLY, 0);
-#endif
-
-            if (m_fd < 0) SCP_FATAL("elf_reader") << "cannot open elf file " << filename();
-
-            Elf* elf = elf_begin(m_fd, ELF_C_READ, nullptr);
-            if (elf == nullptr) SCP_FATAL("elf_reader") << "error reading " << filename() << " : " << elf_errmsg(-1);
-
-            if (elf_kind(elf) != ELF_K_ELF) SCP_FATAL("elf_reader") << "ELF version error in " << filename();
-
-            Elf32_Ehdr* ehdr32 = elf32_getehdr(elf);
-            Elf64_Ehdr* ehdr64 = elf64_getehdr(elf);
-            if (ehdr32) {
-                m_entry = ehdr32->e_entry;
-                m_machine = ehdr32->e_machine;
-                m_endian = elf_endianess(elf);
-                m_segments = elf_segments<elf32_traits>(elf);
-                read_sections<elf32_traits>(elf);
-            }
-
-            if (ehdr64) {
-                m_entry = ehdr64->e_entry;
-                m_machine = ehdr64->e_machine;
-                m_endian = elf_endianess(elf);
-                m_segments = elf_segments<elf64_traits>(elf);
-                read_sections<elf64_traits>(elf);
-            }
-
-            elf_end(elf);
-
-            for (auto s : m_segments) {
-                read_segment(s);
-            }
-        }
-
-        ~elf_reader()
-        {
-            if (m_fd >= 0) close(m_fd);
-        }
-
-        uint64_t read_segment(const elf_segment& segment)
-        {
-            if (m_fd < 0) SCP_FATAL("elf_reader") << "ELF file '" << filename() << "' not open";
-
-            if (lseek(m_fd, segment.offset, SEEK_SET) != (ssize_t)segment.offset)
-                SCP_FATAL("elf_reader") << "cannot seek within ELF file " << filename();
-
-            uint8_t buff[1024];
-            size_t sz = segment.filesz;
-            size_t offset = 0;
-            while (sz) {
-                int s = read(m_fd, buff, (sz < sizeof(buff) ? sz : sizeof(buff)));
-                sz -= s;
-                if (!s) {
-                    SCP_FATAL("elf_reader") << "cannot read ELF file " << filename();
-                    exit(0);
-                }
-                m_send(segment.phys + offset, buff, s);
-                offset += s;
-            }
-            if (lseek(m_fd, 0, SEEK_SET) != 0) SCP_FATAL("elf_reader") << "cannot seek within ELF file " << filename();
-
-            return segment.size;
-        }
-    };
 };
 } // namespace gs
