+++ systemc-components/gs_memory/CMakeLists.txt
@@ -1,7 +1,7 @@
 gs_create_dymod(gs_memory)
 
 target_link_libraries(gs_memory PRIVATE
-    ${LIBELF_LIBRARIES}
+    PkgConfig::LIBELF
 )
 
 if(UNIX AND NOT APPLE)
